DONE: - basic write commit

+++ Comic_app/app/src/main/java/com/example/comic_app/ManageListViewComic.java
@@ -55,7 +55,7 @@ public abstract class ManageListViewComic extends Fragment{
                         comicBook.setId(document.getId());
                         comicBook.setChapterList((List<String>)document.get("chapterList"));
                         comicBook.setAuthor((String)document.get("author"));
-                        comicBook.setCategory((List<Integer>)document.get("category"));
+                        comicBook.setCategory((List<Long>)document.get("category"));
                         comicBook.setImage((String)document.get("image"));
                         comicBook.setLength((String)document.get("length"));
                         comicBook.setStatus((String)document.get("status"));
